requirements.txt
# Core dependencies
beir==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
transformers==4.34.0
sentence-transformers==2.2.2
pyserini==0.21.0
datasets==2.14.4
pyyaml==6.0.1
hydra-core==1.3.2

# Optimization and testing
numba==0.57.1
pytest==7.4.2
pytest-benchmark==4.0.0
py-cpuinfo==9.0.0
psutil==5.9.5
memory-profiler==0.61.0

# BM25 implementations
rank-bm25==0.2.2
pyterrier==0.9.0
splade @ git+https://github.com/naver/splade.git

# Monitoring and profiling
wandb==0.15.11  # Optional: for experiment tracking
numba
memory-profiler